2024-08 Some improvements

+++ src/test/kotlin/year2024/Day08Test.kt
@@ -12,18 +12,18 @@ class Day08Test {
         @JvmStatic
         fun data() = cases(
             arrayOf(
-                Arguments.of("year2024/Day08_sample1", 14L, 34L),
-                Arguments.of("year2024/Day08_sample2", 2L, 5L),
-                Arguments.of("year2024/Day08_sample3", 4L, 8L),
-                Arguments.of("year2024/Day08_sample4", 3L, 9L),
+                Arguments.of("year2024/Day08_sample1", 14, 34),
+                Arguments.of("year2024/Day08_sample2", 2, 5),
+                Arguments.of("year2024/Day08_sample3", 4, 8),
+                Arguments.of("year2024/Day08_sample4", 3, 9),
             ),
-            Arguments.of("year2024/Day08_personal", 351L, 1259L)
+            Arguments.of("year2024/Day08_personal", 351, 1259)
         )
     }
 
     @ParameterizedTest(name = "File {0}, part 1 answer = {1}, part 2 answer = {2}")
     @MethodSource("data")
-    fun test(file: String, part1Answer: Long, part2Answer: Long) {
+    fun test(file: String, part1Answer: Int, part2Answer: Int) {
         val contents = readTestFile(file)
         assertThat(Day08().part1(contents)).isEqualTo(part1Answer)
         assertThat(Day08().part2(contents)).isEqualTo(part2Answer)
